refactor: use kebab-case for files
