+++ package.json
@@ -42,6 +42,7 @@
 		"eslint-config-react": "^1.1.7",
 		"eslint-plugin-prettier": "^4.2.1",
 		"eslint-plugin-react": "^7.30.1",
+		"eslint-plugin-unicorn": "^43.0.2",
 		"husky": "^8.0.1",
 		"lint-staged": "^13.0.3",
 		"prettier": "^2.7.1",
@@ -67,7 +68,8 @@
 		],
 		"parser": "@typescript-eslint/parser",
 		"plugins": [
-			"@typescript-eslint"
+			"@typescript-eslint",
+			"unicorn"
 		],
 		"rules": {
 			"@typescript-eslint/no-unused-vars": [
@@ -80,7 +82,13 @@
 			"@typescript-eslint/no-duplicate-imports": [
 				"error"
 			],
-			"new-cap": "off"
+			"new-cap": "off",
+			"unicorn/filename-case": [
+				"error",
+				{
+					"case": "kebabCase"
+				}
+			]
 		},
 		"settings": {
 			"react": {
