+++ src/pages/user-wallet/public.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { QRCodeSVG } from 'qrcode.react'
 
 // Components
-import { ButtonClose } from '../../components/ButtonClose'
+import { ButtonClose } from '../../components/button-close'
 import { PasswordModal } from '../../components/modals/password/password'
 import { CopyLink } from '../../components/copy-link/copy-link'
 import { Redirect } from '../../components/redirect'
